+++ src/manager/redpepper/manager/config.py
@@ -12,7 +12,7 @@ defaults = {
     "api_logins": [],
     "api_session_secret_key": None,
     "api_session_max_age": 43200,
-    "api_static_dir": "/opt/redpepper/redpepper_console/dist",
+    "api_static_dir": None,
     "api_tls_cert_file": "/etc/redpepper/api.pem",
     "api_tls_key_file": "/etc/redpepper/api-key.pem",
     "api_tls_key_password": None,